validation: limit dynamic indexes to references to matrices and arrays
https://github.com/gpuweb/gpuweb/pull/1801
indexes must be of type 'i32' or 'u32'

Bug: tint:867
Reviewed-on: https://dawn-review.googlesource.com/c/tint/+/54720
Kokoro: Kokoro <noreply+kokoro@google.com>
Auto-Submit: Sarah Mashayekhi <sarahmashay@google.com>

+++ src/transform/bound_array_accessors_test.cc
@@ -52,7 +52,7 @@ TEST_F(BoundArrayAccessorsTest, Array_Idx_Nested_Scalar) {
   auto* src = R"(
 var<private> a : array<f32, 3>;
 
-var<private> b : array<f32, 5>;
+var<private> b : array<i32, 5>;
 
 var<private> i : u32;
 
@@ -64,7 +64,7 @@ fn f() {
   auto* expect = R"(
 var<private> a : array<f32, 3>;
 
-var<private> b : array<f32, 5>;
+var<private> b : array<i32, 5>;
 
 var<private> i : u32;
